requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
bcrypt==4.2.1
cffi==1.17.1
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
psycopg==3.2.3
pycparser==2.22
python-decouple==3.8
sqlparse==0.5.3
tzdata==2024.2